VKT(Frontend): reservation renew test fix
